Add "validating the sign-in form" (10.9)
Seems okay. Added some somewhat believable schema requirements for testing
purposes.

It looks like the Yup error message is only for one
error at a time (no array, though I'm not entirely sure).
This is nice as it's a printable string,
but I guess it might also be nice to show all the issues at once.