Add "environment variables" (10.12)
Seems to work, but the Expo cache seems to be a bit of a pain,
or I guess it's something to do with the Metro bundler? At least doing
"npx expo config" shows the config as updated, but the old content is
still used when logging in the app unless I --clear, in which
case it talks about the bundler cache being empty or whatever.
Annoying, as --clear takes a while to bundle with. Of course,
shouldn't need to change the env too often, but still.