Add "storing the access token step1" (10.14)
It's there and I tested that at least the setting and getting
works.

I've got to say, I'm not entirely sure about the choice to use
classes now. It very much seems like the authStorage would
be a good candidate for a singleton seeing as you'll probably
not be keeping track of authentication for a number users
all at the same time. Additionally, right now it doesn't really
have much in the way of anything meaningful as far as actual state
goes -- though I guess that could change -- so it could quite happily
be done in a functional form too.

Overall, there's the namespace and the key itself, with usually
probably only the set, get, and remove used. I would think
it not too difficult to just define the schema of the storage
as an object:

{
    auth: {
        accessToken: accessTokenSchema
        someOtherAuthNamespaceThing: schemaForTheThing
    }
}

The namespace is the top-level prop, the key is the stringified
prop under the namespace, the schema is used to parse the input,
and something for the actual storage interface? maybe .set, .get,
and .remove, passing the key (which is verified against the above schema,
such that "auth.acesToken" does not work as key), and then pass
the contents for .set. Not obviously much different from the base
functionality, though.

Ooh, or maybe do a URL type situation,
similar to how Express routers are defined, so there'd still be some
modularity if needed:

// the passed functions work as "middleware" to do something before/after
// if a method is left undefined here, it would use a default implementation
authRouter.route("accessTokenKey")
    .get(val => { console.log("Got", val); return val })
    .set(val => return accessTokenSchema.parse(val))
    .remove(val => console.log("Removed value", val))

// then usage
authRouter.set("accessTokenKey", "1234")

Basically it just seems kind of ridiculous to have to define the authStorage
as the functionality that it has would be practically exactly the same
for quite a few different saved things, most likely.