Add "testing the sign in form" (10.18)
The tests pass, but it does occasionally complain about having to
wrap something in act() (something about Formik, Buttons, Animated(View)).
No way of wrapping anything that would sensibly need to be wrapped anyway
actually appears to make any difference, so I just left it as-is.

Now that I think about it, the most obvious culprit is the pressed
Login button, but again, no way of wrapping that seemed to help.

Supposedly the render should be already wrapped by act(), and waitFor
obviously (or similar functionality),
though don't know about since when that has been the case.