Add "repository's review list" (10.20)
Went for a slightly different look, but almost the same. trying to make it
look okay on phone and on the bigger screen too, but with the main focus
naturally being mobile devices, such that the web view can look a little
goofy as I'd assume no-one's going to have a mobile device this big.

Always having to just try all the million centerings as well, and then it's some
textAlignVertical which isn't even a thing in CSS?