Change to using index with picker (10.)
I realised that I wanted to utilise the "text" property of the currently chosen
picker data value, but only the current "value" property was saved in the state.
Of course, I could have compared the current value against each one in
the data array to find the current choice, but that's quite a lot less
efficient than just saving the index of the current choice instead.
Additionally, this sort of prevents a data duplication situation,
as now the "value" property is not unnecessarily in two places, a situation
that would have been even worse had I saved the entire chosen object instead.