Add "infinite scrolling for the repository's reviews list" (10.27)
Seems to be working, mainly just needed a little fiddling with
how the flatlist end reaching works. Would probably also add a separate
refresh button for the end of the list in case the end reaching doesn't
work for whatever reason, such as the view being too short or something.